+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.7.2] - 2026-04-06
+
 ### Added
 
 - `access_mode` parameter to control whether software or hardware access permissions are used.